// sarah.client/src/app/services/api-client/model/namedLocationDto.ts
/**
 * Sarah.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NamedLocationDto { 
    longitude?: number;
    latitude?: number;
    name?: string | null;
}